+++ src/tests/youtube/youtube.h
@@ -40,7 +40,7 @@ struct YoutubeTiming {
     uint64_t stall_count;
     uint64_t total_time;
     uint64_t pre_time;
-    double reported_duration;
+    uint64_t reported_duration;
 };
